% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row-functions.R
\name{chisq_row}
\alias{chisq_row}
\title{Row using chi squared test}
\usage{
chisq_row(
  data_item,
  data = NULL,
  data_filter = NULL,
  row_digits = NULL,
  na.rm = TRUE,
  reference_level = NULL,
  include_reference = TRUE,
  include_denom = NULL,
  percent_first = NULL,
  cat_out_of_row = NULL
)
}
\arguments{
\item{data_item}{item to be taken from data for row}

\item{data}{separate dataset to use}

\item{data_filter}{filter to apply to dataset}

\item{row_digits}{digits for data item (overrides table as a whole)}

\item{na.rm}{whether to include NA in the denominator for percentages}

\item{reference_level}{a level of the variable to drop from display}

\item{include_reference}{whether to include the first level of the factor
in the report}
}
\description{
Row using chi squared test
}
\examples{
first_table(
  mtcars,
  .column_variable = cyl,
  chisq_row(gear, row_digits = 2, include_reference = TRUE)
)
}
