% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row-functions.R
\name{cor_row}
\alias{cor_row}
\title{Correlation row}
\usage{
cor_row(
  data_item,
  data = NULL,
  data_filter = NULL,
  row_digits = NULL,
  method = c("pearson", "kendall", "spearman")
)
}
\arguments{
\item{data_item}{item to be taken from data for row}

\item{data}{separate dataset to use}

\item{data_filter}{filter to apply to dataset}

\item{row_digits}{digits for data item (overrides table as a whole)}

\item{method}{method parameter passed onto \code{\link[stats]{cor}}:
\code{"pearson"}, \code{"kendall"} or \code{"spearman"}.}
}
\description{
Correlation row
}
\examples{
first_table(
  mtcars,
  .column_variable = gear,
  .column_type = "numeric",
  cor_row(disp, method = "spearman")
)
}
