% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row-functions.R
\name{coxph_row}
\alias{coxph_row}
\title{Cox Proportional Hazards Row}
\usage{
coxph_row(data_item, data = NULL, data_filter = NULL, row_digits = NULL,
  include_reference = TRUE)
}
\arguments{
\item{data_item}{item to be taken from data for row}

\item{data}{separate dataset to use}

\item{data_filter}{filter to apply to dataset}

\item{row_digits}{Number of digits to include in the OR}

\item{include_reference}{whether to include a row for the reference level of
a factor}
}
\value{
row for inclusion in `table_one`
}
\description{
Cox Proportional Hazards Row
}
