% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-table.R
\docType{data}
\name{first_table_default_options}
\alias{first_table_default_options}
\title{First Table default options}
\format{An object of class \code{list} of length 8.}
\usage{
first_table_default_options
}
\description{
First Table default options
}
\keyword{datasets}
