% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-table.R
\name{first_table_options}
\alias{first_table_options}
\title{First Table options}
\usage{
first_table_options(digits = 1, include_p = TRUE, p_digits = 3,
  small_p_format = c("<", "E", "x10", "html"), small_p_cutoff = NULL,
  include_n = FALSE, include_n_per_col = c("no", "row", "embed"),
  workspace = 2e+05, default_non_parametric = TRUE, na_text = "NA",
  pretty_p = TRUE, escape_name = TRUE, hide_single_level = FALSE,
  cor_method = c("pearson", "kendall", "spearman"), digits_percent = digits,
  include_denom = FALSE, percent_first = FALSE)
}
\arguments{
\item{digits}{digits used for formatting variables by default}

\item{include_p}{digits used for formatting p values by default}

\item{p_digits}{whether to include p values in table}

\item{small_p_format}{format for small p values}

\item{small_p_cutoff}{cutoff for small p values}

\item{include_n}{whether to include number of non-missing values for each row}

\item{include_n_per_col}{whether to include the number of individuals in each column}

\item{workspace}{default workspace passed onto \code{\link[stats]{fisher.test}}}

\item{default_non_parametric}{whether to default to non parametric tests for
continuous variables}

\item{na_text}{text to use for NA values}

\item{pretty_p}{whether to format p values for display}

\item{escape_name}{whether to escape the row name when displayed as HTML}

\item{hide_single_level}{whether to hide levels for factors when only one}

\item{cor_method}{default correlation method for \code{\link{cor_row}}}

\item{digits_percent}{digits used by default for percentages
(overrides \code{digits})}

\item{include_denom}{whether to include the denominator for categorical
variables}

\item{percent_first}{whether to put the percent before the n for categorical
variables}
}
\description{
First Table options
}
