% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row-functions.R
\name{first_table_row}
\alias{first_table_row}
\title{Row with type selected by firsttable}
\usage{
first_table_row(
  data_item,
  data = NULL,
  data_filter = NULL,
  row_digits = NULL,
  na.rm = TRUE,
  reference_level = NULL,
  include_reference = NULL,
  workspace = NULL,
  non_parametric = NULL,
  row_digits_default = NULL,
  row_digits_surv = NULL,
  row_digits_numeric = NULL,
  cat_out_of_row = NULL
)
}
\arguments{
\item{data_item}{item to be taken from data for row}

\item{data}{separate dataset to use}

\item{data_filter}{filter to apply to dataset}

\item{row_digits}{digits for data item (overrides table as a whole)}

\item{na.rm}{whether to remove NA before reporting median and quartiles}

\item{reference_level}{a level of the variable to drop from display (only
relevant for logical/factor/character variables)}

\item{include_reference}{whether to include a row for the reference level of
a factor (only relevant for logical/factor/character variables)}

\item{workspace}{passed onto \code{\link[stats]{fisher.test}}}

\item{non_parametric}{whether to use non-parametric tests}

\item{row_digits_default}{digits where \code{.column_type = "default"}}

\item{row_digits_surv}{digits where \code{.column_type = "default"} and
\code{.column_variable} inherits \code{Surv}}

\item{row_digits_numeric}{digits where \code{.column_type = "numeric"}}

\item{cat_out_of_row}{whether percentages in categories should be calculated
out of row rather than column}
}
\value{
row for inclusion in `first_table`
}
\description{
Row with type selected by firsttable
}
\details{
This provides a generic row for \code{\link{first_table}} with
the type of row determined from the \code{class} of the data. This allows a
\code{list} of \code{\link[rlang]{quos}} to be created and then used for
both a standard \code{\link{first_table}} and one that uses a
\code{\link[survival]{Surv}} column.
}
\examples{
library(survival)
first_table(lung,
  .column_variable = Surv(time, status),
  .options = list(include_n = TRUE, include_n_per_col = TRUE),
   `Meal calories` = first_table_row(meal.cal, row_digits = 2)
 )
}
