% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row-functions.R
\name{fisher_row}
\alias{fisher_row}
\title{Title}
\usage{
fisher_row(data_item, row_digits = NULL, row_p_digits = NULL,
  na.rm = TRUE)
}
\arguments{
\item{data_item}{item to be taken from data for row}

\item{row_digits}{digits for data item (overrides table as a whole)}

\item{row_p_digits}{digits for p value (overrides table as a whole)}

\item{na.rm}{whether to include NA in the denominator for percentages}
}
\description{
Title
}
