% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row-functions.R
\name{kruskal_row}
\alias{kruskal_row}
\title{Kruskal Wallis test row}
\usage{
kruskal_row(data_item, data = NULL, data_filter = NULL, row_digits = NULL,
  na.rm = TRUE)
}
\arguments{
\item{data_item}{item to be taken from data for row}

\item{data}{separate dataset to use}

\item{data_filter}{filter to apply to dataset}

\item{row_digits}{digits for data item (overrides table as a whole)}

\item{na.rm}{whether to remove NA before reporting median and quartiles}
}
\description{
Kruskal Wallis test row
}
\examples{
first_table(
  mtcars,
  .column_variable = cyl,
  kruskal_row(disp, row_digits = 2)
)
}
