% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row-functions.R
\name{pretty_p}
\alias{pretty_p}
\title{Format p values for display}
\usage{
pretty_p(p, p_digits, small_p_format = c("<", "E", "x10", "plotmath", "html"),
  small_p_cutoff = 10^-p_digits)
}
\arguments{
\item{p}{p value to format}

\item{p_digits}{number of digits to display}

\item{small_p_format}{format used to display p values below a threshold}

\item{small_p_cutoff}{cutoff for displaying alternative formatting}
}
\value{
formatted p value
}
\description{
Format p values for display
}
