% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-one.R
\name{table_one}
\alias{table_one}
\title{Table One}
\usage{
table_one(data, ..., column_variable = NULL, digits = 1, p_digits = 3,
  small_p = c("<", "E", "x10"))
}
\arguments{
\item{data}{`data.frame` or `tibble` to use as data source}

\item{...}{row details}

\item{column_variable}{variable used for columns (if any)}

\item{digits}{digits used for formatting variables by default}

\item{p_digits}{digits used for formatting p values by default}

\item{small_p}{format used for small p values}
}
\value{
character matrix with table
}
\description{
Table One
}
