% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-table.R
\name{first_table}
\alias{first_table}
\title{First Table}
\usage{
first_table(data, ..., column_variable = NULL, digits = 1,
  include_p = TRUE, p_digits = 3, small_p_format = c("<", "E", "x10",
  "html"), small_p_cutoff = 10^-p_digits, include_n = FALSE,
  include_n_per_col = FALSE, workspace = 2e+05)
}
\arguments{
\item{data}{`data.frame` or `tibble` to use as data source}

\item{...}{row details}

\item{column_variable}{variable used for columns (if any)}

\item{digits}{digits used for formatting variables by default}

\item{include_p}{whether to include p values in table}

\item{p_digits}{digits used for formatting p values by default}

\item{small_p_format}{format for small p values}

\item{small_p_cutoff}{cutoff for small p values}

\item{include_n}{whether to include number of non-missing values for each row}

\item{include_n_per_col}{whether to include the number of individuals in each column}

\item{workspace}{default workspace passed onto \code{\link[stats]{fisher.test}}}
}
\value{
character matrix with the requested rows and columns
}
\description{
First Table
}
\details{
This function takes a \code{\link[base]{data.frame}} or \code{\link[tibble]{tibble}}
and a row and column specification and generates a table along the lines of the first table
used in many medical journal articles. The row specification can either use one of the \code{_row}
functions, or if the defaults are appropriate can just be a bare column name or calculation
using a column. These calculations are implemented using \code{\link[rlang]{eval_tidy}} and
support the \code{\link[rlang]{quasiquotation}} operators such as \code{\link[rlang]{!!}} and
\code{\link[rlang]{!!!}}.

The output is a character matrix which is formatted so as to work well with \code{\link[pander]{pander}}
and \code{\link[knitr]{knit}}.
}
\examples{
first_table(mtcars,
  column_variable = cyl,
  "Miles per gallon" = mpg,
  "Transmission" = factor(am))

}
