% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-table.R
\name{first_table}
\alias{first_table}
\title{First Table}
\usage{
first_table(data, ..., .column_variable = NULL,
  .options = first_table_default_options)
}
\arguments{
\item{data}{`data.frame` or `tibble` to use as data source}

\item{...}{row details}

\item{.options}{options to use for formatting (see details)}

\item{column_variable}{variable used for columns (if any)}
}
\value{
character matrix with the requested rows and columns
}
\description{
First Table
}
\details{
This function takes a \code{\link[base]{data.frame}} or \code{\link[tibble]{tibble}}
and a row and column specification and generates a table along the lines of the first table
used in many medical journal articles. The row specification can either use one of the \code{_row}
functions, or if the defaults are appropriate can just be a bare column name or calculation
using a column. These calculations are implemented using \code{\link[rlang]{eval_tidy}} and
support the \code{\link[rlang]{quasiquotation}} operators such as \code{\link[rlang]{!!}} and
\code{\link[rlang]{!!!}}.

Options are currently:
\code{digits} digits used for formatting variables by default
\code{p_digits} digits used for formatting p values by default
\code{include_p} whether to include p values in table
\code{small_p_format} format for small p values
\code{small_p_cutoff} cutoff for small p values
\code{include_n} whether to include number of non-missing values for each row
\code{include_n_per_col} whether to include the number of individuals in each column
\code{workspace} default workspace passed onto \code{\link[stats]{fisher.test}}

The output is a character matrix which is formatted so as to work well with \code{\link[pander]{pander}}
and \code{\link[knitr]{knit}}.
}
\examples{
first_table(mtcars,
  .column_variable = cyl,
  "Miles per gallon" = mpg,
  "Transmission" = factor(am))

}
