% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-table.R
\name{first_table_options}
\alias{first_table_options}
\title{First Table options}
\usage{
first_table_options(digits = 1, include_p = TRUE, p_digits = 3,
  small_p_format = c("<", "E", "x10", "html"), small_p_cutoff = NULL,
  include_n = FALSE, include_n_per_col = FALSE, workspace = 2e+05,
  default_non_parametric = TRUE)
}
\arguments{
\item{digits}{digits used for formatting variables by default}

\item{include_p}{digits used for formatting p values by default}

\item{p_digits}{whether to include p values in table}

\item{small_p_format}{format for small p values}

\item{small_p_cutoff}{cutoff for small p values}

\item{include_n}{whether to include number of non-missing values for each row}

\item{include_n_per_col}{whether to include the number of individuals in each column}

\item{workspace}{default workspace passed onto \code{\link[stats]{fisher.test}}}

\item{default_non_parametric}{whether to default to non parametric tests for
continuous variables (note only applies to rows without an explicit row function)}
}
\description{
First Table options
}
