% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-table.R
\name{first_table_options}
\alias{first_table_options}
\title{First Table options}
\usage{
first_table_options(
  template = NULL,
  digits = 1,
  include_p = TRUE,
  p_digits = 3,
  small_p_format = c("<", "E", "x10", "html"),
  small_p_cutoff = NULL,
  include_n = FALSE,
  include_n_per_col = c("no", "row", "embed"),
  workspace = 2e+05,
  default_non_parametric = TRUE,
  na_text = "NA",
  pretty_p = TRUE,
  p_sig_fig = FALSE,
  p_n_sig_fig = 2,
  escape_name = TRUE,
  hide_single_level = FALSE,
  cor_method = c("pearson", "kendall", "spearman"),
  digits_percent = digits,
  include_denom = FALSE,
  percent_first = FALSE,
  hybrid_fisher = FALSE,
  simulate_p_value_fisher = FALSE,
  include_estimate_diff = FALSE,
  factor_name_own_row = FALSE,
  cat_out_of_row = FALSE,
  include_overall_column = FALSE,
  hide_level_logical = FALSE,
  use_interpuncts = FALSE
)
}
\arguments{
\item{template}{existing first_table_options to populate defaults}

\item{digits}{digits used for formatting variables by default}

\item{include_p}{digits used for formatting p values by default}

\item{p_digits}{whether to include p values in table}

\item{small_p_format}{format for small p values}

\item{small_p_cutoff}{cutoff for small p values}

\item{include_n}{whether to include number of non-missing values for each row}

\item{include_n_per_col}{whether to include the number of individuals in each column}

\item{workspace}{default workspace passed onto \code{\link[stats]{fisher.test}}}

\item{default_non_parametric}{whether to default to non parametric tests for
continuous variables}

\item{na_text}{text to use for NA values}

\item{pretty_p}{whether to format p values for display}

\item{p_sig_fig}{whether to use significant figures for p value (rather than
decimal digits)}

\item{p_n_sig_fig}{number of significant figures}

\item{escape_name}{whether to escape the row name when displayed as HTML}

\item{hide_single_level}{whether to hide levels for factors when only one}

\item{cor_method}{default correlation method for \code{\link{cor_row}}}

\item{digits_percent}{digits used by default for percentages
(overrides \code{digits})}

\item{include_denom}{whether to include the denominator for categorical
variables}

\item{percent_first}{whether to put the percent before the n for categorical
variables}

\item{hybrid_fisher}{whether to use a hybrid approach for
\code{\link[stats]{fisher.test}} and >2x2 tables}

\item{simulate_p_value_fisher}{whether to simulate p-values for
\code{\link[stats]{fisher.test}} and >2x2 tables}

\item{include_estimate_diff}{whether to include an estimate of the difference
for continuous data (with appropriate estimates for parametric and
non-parametric data)}

\item{factor_name_own_row}{whether to have the name of a factor in a row
on its own; only affects huxtable output}

\item{cat_out_of_row}{whether percentages in categories should be calculated
out of row rather than column}

\item{include_overall_column}{whether to include an overall column in
addition to separate columns by column variable}

\item{hide_level_logical}{hide the display of the level TRUE for logical rows}

\item{use_interpuncts}{replaces decimal points with interpuncts;
most commonly used for Lancet journals}
}
\description{
First Table options
}
