% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row-functions.R
\name{first_table_row}
\alias{first_table_row}
\title{Row with type selected by firsttable}
\usage{
first_table_row(data_item, data = NULL, data_filter = NULL,
  row_digits = NULL, na.rm = TRUE, reference_level = NULL,
  include_reference = NULL, workspace = 2e+05)
}
\arguments{
\item{data_item}{item to be taken from data for row}

\item{data}{separate dataset to use}

\item{data_filter}{filter to apply to dataset}

\item{row_digits}{digits for data item (overrides table as a whole)}

\item{na.rm}{whether to remove NA before reporting median and quartiles}

\item{include_reference}{whether to include a row for the reference level of
a factor}

\item{workspace}{passed onto \code{\link[stats]{fisher.test}}}

\item{include_reference}{whether to include the first level of the factor
in the report}
}
\description{
Row with type selected by firsttable
}
