% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row-functions.R
\name{fisher_row}
\alias{fisher_row}
\title{Row using Fisher's exact test}
\usage{
fisher_row(
  data_item,
  data = NULL,
  data_filter = NULL,
  row_digits = NULL,
  na.rm = TRUE,
  reference_level = NULL,
  include_reference = TRUE,
  workspace = NULL,
  include_denom = NULL,
  percent_first = NULL,
  cat_out_of_row = NULL
)
}
\arguments{
\item{data_item}{item to be taken from data for row}

\item{data}{separate dataset to use}

\item{data_filter}{filter to apply to dataset}

\item{row_digits}{digits for data item (overrides table as a whole)}

\item{na.rm}{whether to include NA in the denominator for percentages}

\item{reference_level}{a level of the variable to drop from display}

\item{include_reference}{whether to include the first level of the factor
in the report}

\item{workspace}{passed onto \code{\link[stats]{fisher.test}}}

\item{include_denom}{whether to include the denominator for categorical
variables}

\item{percent_first}{whether to put the percent before the n for categorical
variables}
}
\description{
Row using Fisher's exact test
}
\examples{
first_table(
  mtcars,
  .column_variable = cyl,
  fisher_row(gear, row_digits = 2, include_reference = TRUE)
)
}
