% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prety-p.R
\name{pretty_p}
\alias{pretty_p}
\title{Format p values for display}
\usage{
pretty_p(p, p_digits, small_p_format = c("<", "E", "x10", "plotmath", "html"),
  small_p_cutoff = 10^-p_digits)
}
\arguments{
\item{p}{p value to format}

\item{p_digits}{number of digits to display}

\item{small_p_format}{format used to display p values below a threshold}

\item{small_p_cutoff}{cutoff for displaying alternative formatting}
}
\value{
formatted p value
}
\description{
Format p values for display
}
\examples{
p_values <- 10 ^ seq(-1, -6, -1)
# Format p values as for example 1.0x10^-4
pretty_p(p_values, p_digits = 3, small_p_format = "x10")
# Format p values for use on a graph
x <- rlnorm(100, 3, 1)
y <- rlnorm(100, 4, 1)
p <- wilcox.test(x, y)$p.value
p_formatted <- pretty_p(p, p_digits = 3, small_p_format = "plotmath")
boxplot(x, y, ylim = c(0, max(c(x, y)) + 100))
text(1.5, max(c(x, y) + 50), parse(text = paste0("p == ", p_formatted)))
}
