% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-one.R
\name{table_one}
\alias{table_one}
\title{Table One}
\usage{
table_one(data, ..., column_variable = NULL, digits = 1, include_p = TRUE,
  p_digits = 3, small_p_format = c("<", "E", "x10", "html"),
  small_p_cutoff = 10^-p_digits, include_n = FALSE)
}
\arguments{
\item{data}{`data.frame` or `tibble` to use as data source}

\item{...}{row details}

\item{column_variable}{variable used for columns (if any)}

\item{digits}{digits used for formatting variables by default}

\item{include_p}{whether to include p values in table}

\item{p_digits}{digits used for formatting p values by default}

\item{small_p_format}{format for small p values}

\item{small_p_cutoff}{cutoff for small p values}

\item{include_n}{whether to include number of non-missing values for each row}
}
\value{
character matrix with table
}
\description{
Table One
}
